+++ modules/sde-submodules/part-as-planned/src/main/java/org/eclipse/tractusx/sde/submodels/pap/steps/EDCPartAsPlannedHandlerStep.java
@@ -19,6 +19,7 @@
  ********************************************************************************/
 package org.eclipse.tractusx.sde.submodels.pap.steps;
 
+import java.util.List;
 import java.util.Map;
 
 import org.eclipse.tractusx.sde.common.constants.CommonConstants;
@@ -88,7 +89,7 @@ public class EDCPartAsPlannedHandlerStep extends Step {
 	private void edcProcessingforPartAsPlanned(AssetEntryRequest assetEntryRequest, PartAsPlanned input) {
 
 		Map<String, String> createEDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest,
-				input.getBpnNumbers(), input.getUsagePolicies());
+				List.of(), List.of());
 
 		// EDC transaction information for DB
 		input.setAssetId(assetEntryRequest.getId());