+++ modules/sde-submodules/serial-part-typization/src/main/java/org/eclipse/tractusx/sde/submodels/spt/steps/EDCAspectHandlerUseCase.java
@@ -22,6 +22,7 @@
 
 package org.eclipse.tractusx.sde.submodels.spt.steps;
 
+import java.util.List;
 import java.util.Map;
 
 import org.eclipse.tractusx.sde.common.constants.CommonConstants;
@@ -92,7 +93,7 @@ public class EDCAspectHandlerUseCase extends Step {
 	private void edcProcessingforAspect(AssetEntryRequest assetEntryRequest, Aspect input) {
 
 		Map<String, String> createEDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest,
-				input.getBpnNumbers(), input.getUsagePolicies());
+				List.of(),List.of());
 
 		// EDC transaction information for DB
 		input.setAssetId(assetEntryRequest.getId());